% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas-package.R
\docType{package}
\name{TWAS-package}
\alias{TWAS-package}
\title{TWAS: A package for transcriptome-wide association studies}
\description{
The TWAS package provides a single function, \code{twas}, for
jointly modeling two independent data sets: trait data (genome-wide
association, GWA) and expression data (transcriptome-wide
association, TWA). The analytical goal is to leverage the
transcriptome data in analysis of the GWA data to inform the
mechanisms underlying gene-trait associations.
}
