% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twas.gene.markers}
\alias{twas.gene.markers}
\title{Sample sets of markers proximal to genes}
\format{An object of class \code{list} of length 4.}
\usage{
twas.gene.markers
}
\description{
\code{twas.gene.markers} is a list of character vectors of marker
names for four sample gene names. The gene symbol is the name of
each character vector.
}
\keyword{datasets}
