% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twas.traits}
\alias{twas.traits}
\title{Sample trait data for package \code{twas}.}
\format{traits - A data.frame of 174 rows with one ID column and 2 traits
\describe{
  \item{ID}{subject identifier}
  \item{logCRP}{log CRP}
  \item{logIL6}{log IL6}
}

genotypes - A data.frame of 174 rows with one ID column and genotypes for 30 markers
\describe{
  \item{ID}{subject identifier}
  \item{rs13392501}{genotypes for marker rs13392501}
  \item{etc...}
}}
\usage{
twas.traits
}
\description{
\code{twas.traits} is a list containing two data.frames as follows:
}
\keyword{datasets}
