% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twas.tx}
\alias{twas.tx}
\title{Sample transcriptome data for package \code{twas}.}
\format{transcriptome - A data.frame of 922 rows with one ID column and 4 columns of gene expression:
\describe{
  \item{ID}{subject identifier}
  \item{COL5A2}{COL5A2 RNASeq transcript counts}
  \item{LRRC16A}{LRRC16A RNASeq counts}
  \item{SLC9A8}{SLC9A8 RNASeq counts}
  \item{PDIA5}{PDIA5 RNASeq counts}
}

genotypes - A data.frame of 922 rows with one ID column and genotypes for 30 markers
\describe{
  \item{ID}{subject identifier}
  \item{rs13392501}{genotypes for marker rs13392501}
  \item{etc...}
}}
\usage{
twas.tx
}
\description{
\code{twas.tx} is a list containing two data.frames as follows:
}
\keyword{datasets}
